+++ packages/core/runtime/stores/storeDesktopVolume.ts
@@ -1,8 +1,9 @@
 import { defineStore } from 'pinia'
 import { computed } from 'vue'
 import { useDesktopStore } from './storeDesktop'
+import { DESKTOP_VOLUME_STORE_ID } from './storeIds'
 
-export const useDesktopVolumeStore = defineStore('owd/desktop/volume', () => {
+export const useDesktopVolumeStore = defineStore(DESKTOP_VOLUME_STORE_ID, () => {
   const desktopStore = useDesktopStore()
 
   const master = computed(() => desktopStore.state.volume.master)